Development

+++ src/shared/unlinkFile.ts
@@ -1,11 +1,14 @@
-import fs from 'fs';
-import path from 'path';
+import fs from "fs";
+import path from "path";
 
 const unlinkFile = (file: string) => {
-    const filePath = path.join('uploads', file);
-    if (fs.existsSync(filePath)) {
-        fs.unlinkSync(filePath);
-    }
+  if (!file) {
+    return;
+  }
+  const filePath = path.join("uploads", file);
+  if (fs.existsSync(filePath)) {
+    fs.unlinkSync(filePath);
+  }
 };
 
 export default unlinkFile;
